cuga-apps/apps/deck_forge/requirements.txt
fastapi
uvicorn[standard]
pydantic
pdfplumber
python-pptx
chromadb
sentence-transformers
langchain-core
langgraph
faster-whisper
httpx
